+++ CHANGELOG.md
@@ -59,6 +59,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
    * Update minimum numpy in CI tests to 1.18 following NEP29
    * Made `InstTestClass` more portable to streamline user implementation of the
      standard end-to-end instrument tests.
+   * Moved test classes to `pysat.tests.classes`
+   * Reorganized the core test_instrument.py as a series of classes
 
 [3.0.1] - 2021-07-28
 --------------------